Moved IRequestNotify to application project
