+++ src/HttPlaceholder/Hubs/Implementations/RequestNotify.cs
@@ -1,4 +1,7 @@
 ﻿using System.Threading.Tasks;
+using AutoMapper;
+using HttPlaceholder.Application.Interfaces.Signalling;
+using HttPlaceholder.Domain;
 using HttPlaceholder.Dto.v1.Requests;
 using Microsoft.AspNetCore.SignalR;
 
@@ -8,15 +11,21 @@ namespace HttPlaceholder.Hubs.Implementations;
 public class RequestNotify : IRequestNotify
 {
     private readonly IHubContext<RequestHub> _hubContext;
+    private readonly IMapper _mapper;
 
     /// <summary>
     /// Constructs a <see cref="RequestNotify"/> instance.
     /// </summary>
-    public RequestNotify(IHubContext<RequestHub> hubContext)
+    public RequestNotify(IHubContext<RequestHub> hubContext, IMapper mapper)
     {
         _hubContext = hubContext;
+        _mapper = mapper;
     }
 
     /// <inheritdoc />
-    public async Task NewRequestReceivedAsync(RequestOverviewDto request) => await _hubContext.Clients.All.SendAsync("RequestReceived", request);
+    public async Task NewRequestReceivedAsync(RequestResultModel request)
+    {
+        var input = _mapper.Map<RequestOverviewDto>(request);
+        await _hubContext.Clients.All.SendAsync("RequestReceived", input);
+    }
 }
